requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
appnope==0.1.2
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==8.0.3
cryptography==3.4.7
cycler==0.10.0
decorator==5.1.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.33.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.37.1
h5py==3.1.0
idna==3.2
importlib-metadata==4.8.1
ipython==7.26.0
jedi==0.18.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.2
msgpack==1.0.2
multidict==5.1.0
numpy==1.21.4
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.4.58
opt-einsum==3.3.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pip==21.3.1
prompt-toolkit==3.0.20
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pynvim==0.4.3
pyOpenSSL==21.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.1
setuptools==58.4.0
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.6.0
tensorflow-addons==0.14.0.dev0
tensorflow-estimator==2.6.0
tensorflow-hub==0.8.0
tensorflow-macos==2.6.0
tensorflow-metal==0.2.0
termcolor==1.1.0
tornado==6.1
traitlets==5.1.0
typeguard==2.13.2
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wheel==0.37.0
wrapt==1.12.1
yarl==1.6.3
zipp==3.6.0
